Make wind_mixing/uw_train_NDE.jl work
